chore: remove chat module
