+++ frappe/__init__.py
@@ -1797,7 +1797,7 @@ def get_version(doctype, name, limit=None, head=False, raise_err=True):
 			'limit': limit
 		}, as_list=1)
 
-		from frappe.chat.util import squashify, dictify, safe_json_loads
+		from frappe.util import squashify, dictify, safe_json_loads
 
 		versions = []
 
@@ -1854,7 +1854,7 @@ def mock(type, size=1, locale='en'):
 			data = getattr(fake, type)()
 			results.append(data)
 
-	from frappe.chat.util import squashify
+	from frappe.util import squashify
 	return squashify(results)
 
 def validate_and_sanitize_search_inputs(fn):
