--- frappe/chat/doctype/chat_message/chat_message_list.js
@@ -1,8 +1,0 @@
-frappe.listview_settings['Chat Message'] = {
-	filters: [
-		['Chat Message', 'user',  '==', frappe.session.user, true]
-		// I need an or_filter here.
-		// ['Chat Room',    'owner', '==', frappe.session.user, true],
-		// ['Chat Room',    frappe.session.user, 'in', 'users', true]
-	]
-};
