--- frappe/chat/doctype/chat_profile/chat_profile_list.js
@@ -1,11 +1,0 @@
-frappe.listview_settings['Chat Profile'] = 
-{
-	get_indicator: function (doc)
-	{
-		const status = frappe.utils.squash(frappe.chat.profile.STATUSES.filter(
-			s => s.name === doc.status
-		));
-
-		return [__(status.name), status.color, `status,=,${status.name}`]
-	}
-};
