--- frappe/chat/doctype/chat_room/chat_room.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2017, Frappe Technologies and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('Chat Room', {
-	refresh: function (form) {
-
-	}
-});
