--- frappe/chat/doctype/chat_room/chat_room_list.js
@@ -1,6 +1,0 @@
-frappe.listview_settings['Chat Room'] = {
-	filters: [
-		['Chat Room', 'owner', '=', frappe.session.user, true],
-		['Chat Room User', 'user', '=', frappe.session.user, true]
-	]
-};
