--- frappe/chat/doctype/chat_room_user/chat_room_user.json
@@ -1,40 +1,0 @@
-{
- "beta": 1,
- "creation": "2017-11-08 15:24:21.029314",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "user",
-  "is_admin"
- ],
- "fields": [
-  {
-   "fieldname": "user",
-   "fieldtype": "Link",
-   "in_list_view": 1,
-   "label": "User",
-   "options": "User",
-   "reqd": 1
-  },
-  {
-   "default": "0",
-   "fieldname": "is_admin",
-   "fieldtype": "Check",
-   "label": "Admin"
-  }
- ],
- "in_create": 1,
- "istable": 1,
- "modified": "2019-11-07 13:21:05.297337",
- "modified_by": "Administrator",
- "module": "Chat",
- "name": "Chat Room User",
- "owner": "Administrator",
- "permissions": [],
- "quick_entry": 1,
- "read_only": 1,
- "sort_field": "modified",
- "sort_order": "DESC",
- "track_changes": 1
-}
