--- frappe/chat/doctype/chat_room_user/chat_room_user.py
@@ -1,10 +1,0 @@
-from __future__ import unicode_literals
-
-# imports - module imports
-from   frappe.model.document import Document
-import frappe
-
-session = frappe.session
-
-class ChatRoomUser(Document):
-	pass
