--- frappe/chat/doctype/chat_token/chat_token.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2018, Frappe Technologies and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('Chat Token', {
-	refresh: function(frm) {
-
-	}
-});
