--- frappe/chat/doctype/chat_token/chat_token.json
@@ -1,57 +1,0 @@
-{
- "autoname": "field:token",
- "beta": 1,
- "creation": "2018-03-26 18:20:13.825652",
- "doctype": "DocType",
- "editable_grid": 1,
- "engine": "InnoDB",
- "field_order": [
-  "token",
-  "ip_address",
-  "country"
- ],
- "fields": [
-  {
-   "fieldname": "token",
-   "fieldtype": "Data",
-   "in_list_view": 1,
-   "label": "Token",
-   "reqd": 1
-  },
-  {
-   "fieldname": "ip_address",
-   "fieldtype": "Data",
-   "label": "IP Address"
-  },
-  {
-   "fieldname": "country",
-   "fieldtype": "Data",
-   "label": "Country"
-  }
- ],
- "in_create": 1,
- "modified": "2019-11-07 13:21:24.514558",
- "modified_by": "Administrator",
- "module": "Chat",
- "name": "Chat Token",
- "owner": "Administrator",
- "permissions": [
-  {
-   "create": 1,
-   "delete": 1,
-   "email": 1,
-   "export": 1,
-   "print": 1,
-   "read": 1,
-   "report": 1,
-   "role": "System Manager",
-   "share": 1,
-   "write": 1
-  }
- ],
- "quick_entry": 1,
- "read_only": 1,
- "sort_field": "modified",
- "sort_order": "DESC",
- "track_changes": 1
-}
