--- frappe/chat/doctype/chat_token/chat_token.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Copyright (c) 2018, Frappe Technologies and contributors
-# For license information, please see license.txt
-
-from __future__ import unicode_literals
-import frappe
-from frappe.model.document import Document
-
-class ChatToken(Document):
-	pass
