--- frappe/chat/util/__init__.py
@@ -1,15 +1,0 @@
-from __future__ import unicode_literals
-
-# imports - module imports
-from frappe.chat.util.util import (
-	get_user_doc,
-	squashify,
-	safe_json_loads,
-	filter_dict,
-	get_if_empty,
-	listify,
-	dictify,
-	check_url,
-	create_test_user,
-	get_emojis
-)
