+++ frappe/core/doctype/user/user.json
@@ -619,11 +619,6 @@
   },
   {
    "group": "Profile",
-   "link_doctype": "Chat Profile",
-   "link_fieldname": "user"
-  },
-  {
-   "group": "Profile",
    "link_doctype": "Blogger",
    "link_fieldname": "user"
   },
