+++ frappe/modules.txt
@@ -9,7 +9,6 @@ Integrations
 Printing
 Contacts
 Data Migration
-Chat
 Social
 Automation
 Event Streaming
