+++ frappe/website/doctype/website_settings/website_settings.py
@@ -122,8 +122,7 @@ def get_website_settings(context=None):
 		"facebook_share", "google_plus_one", "twitter_share", "linked_in_share",
 		"disable_signup", "hide_footer_signup", "head_html", "title_prefix",
 		"navbar_template", "footer_template", "navbar_search", "enable_view_tracking",
-		"footer_logo", "call_to_action", "call_to_action_url", "show_language_picker",
-		"chat_enable"]:
+		"footer_logo", "call_to_action", "call_to_action_url", "show_language_picker"]:
 		if hasattr(settings, k):
 			context[k] = settings.get(k)
